Combine into fuel-indexer-benchmarks
