--- benches/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-edition = { workspace = true }
-license = { workspace = true }
-name = "fuel-indexer-benches"
-publish = false
-version = "0.0.0"
-
-[lib]
-name = "fuel_indexer_benches"
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-criterion = { version = "0.5.1", features = ["html_reports", "async_tokio"] }
-fuel-indexer = { workspace = true }
-fuel-indexer-database = { workspace = true }
-fuel-indexer-graphql = { workspace = true }
-fuel-indexer-lib = { workspace = true }
-fuel-indexer-schema = { workspace = true }
-fuel-indexer-tests = { path = "../packages/fuel-indexer-tests" }
-fuel-core-client = "0.17.2"
-futures = "0.3"
-tokio = { workspace = true, features = ["macros", "rt-multi-thread"]}
-
-[[bench]]
-harness = false
-name = "indexer_examples"
-
-[[bench]]
-harness = false
-name = "graphql"
