--- benches/benches/wasm_examples/assets/default_indexer/schema/default_indexer.schema.graphql
@@ -1,12 +1,0 @@
-type Block {
-    id: ID!
-    height: UInt8!
-    hash: Bytes32! @unique
-}
-
-type Transaction {
-    id: ID!
-    block: Block! @join(on:hash)
-    hash: Bytes32! @unique
-}
-
