--- benches/benches/wasm_examples/mod.rs
@@ -1,2 +1,0 @@
-pub mod default_indexer;
-pub mod fuel_explorer;
