+++ packages/fuel-indexer-benchmarks/benches/graphql.rs
