+++ packages/fuel-indexer-benchmarks/benches/wasm/examples/mod.rs
@@ -1,0 +1,1 @@
+pub mod fuel_explorer;
