+++ packages/fuel-indexer-benchmarks/benches/wasm/metering.rs
@@ -1,0 +1,7 @@
+// use criterion::Criterion;
+
+// use fuel_indexer::{IndexerConfig, Manifest};
+// use fuel_indexer_benchmarks::{create_wasm_indexer_benchmark, create_wasm_manifest};
+
+// TODO: Create a custom measurement that returns the
+// amount of points used by an executor
