+++ scripts/utils/build_test_wasm_module.bash
@@ -29,13 +29,4 @@ bash scripts/stripper.bash hello_indexer.wasm
 cp hello_indexer.wasm target/wasm32-unknown-unknown/release/
 rm -fv hello_indexer.wasm
 
-mkdir -p /tmp/default_indexer
-if [ ! -f "/tmp/default_indexer/Cargo.toml" ]; then
-    forc index init --path /tmp/default_indexer --namespace my_project
-fi
-cargo build --manifest-path /tmp/default_indexer/Cargo.toml --release --target wasm32-unknown-unknown
-wasm-snip /tmp/default_indexer/target/wasm32-unknown-unknown/release/default_indexer.wasm -o default_indexer.wasm -p __wbindgen
-cp default_indexer.wasm target/wasm32-unknown-unknown/release/
-rm -fv default_indexer.wasm
-
 set +ex
